refactor data files and folder
