+++ api/routes/auth-users.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { db } from "../data/database";
+import { db } from "../database/database";
 
 const userRouter = Router()
